+++ packages/moonstone/internal/SliderDecorator/SliderDecorator.js
@@ -172,6 +172,15 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			onKnobMove: PropTypes.func,
 
 			/**
+			 * Sets the orientation of the slider, whether the slider moves left and right or up and
+			 * down. Must be either `'horizontal'` or `'vertical'`.
+			 *
+			 * @type {String}
+			 * @public
+			 */
+			orientation: PropTypes.oneOf(['horizontal', 'vertical']),
+
+			/**
 			 * The amount to increment or decrement the value.
 			 *
 			 * @type {Number}
@@ -195,22 +204,14 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			 * @default 0
 			 * @public
 			 */
-			value: PropTypes.number,
-
-			/**
-			 * If `true` the slider will be oriented vertically.
-			 *
-			 * @type {Boolean}
-			 * @public
-			 */
-			vertical: PropTypes.bool
+			value: PropTypes.number
 		};
 
 		static defaultProps = {
 			max: 100,
 			min: 0,
-			step: 1,
-			vertical: false
+			orientation: 'horizontal',
+			step: 1
 		};
 
 		constructor (props) {
@@ -382,7 +383,8 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		updateUI = () => {
 			// intentionally breaking encapsulation to avoid having to specify multiple refs
 			const {barNode, knobNode, loaderNode, node} = this.sliderBarNode;
-			const {backgroundProgress, vertical} = this.props;
+			const {backgroundProgress, orientation} = this.props;
+			const vertical = orientation === 'vertical';
 			const {value} = this.state;
 			const proportionProgress = computeProportionProgress({value, max: this.normalizedMax, min: this.normalizedMin});
 			const knobProgress = this.knobPosition != null ? this.knobPosition : proportionProgress;
@@ -457,7 +459,7 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		handleMouseEnter = (ev) => {
 			forwardMouseEnter(ev, this.props);
 
-			if (!this.props.detachedKnob || this.props.disabled || this.props.vertical) return;
+			if (!this.props.detachedKnob || this.props.disabled || this.props.orientation === 'vertical') return;
 
 			this.moveKnobByPointer(ev.clientX);
 		}
@@ -466,7 +468,7 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 			forwardMouseMove(ev, this.props);
 			this.willChange = Boolean(ev.buttons); // if detached knob is dragging, it fires onChange
 
-			if (!this.props.detachedKnob || this.props.disabled || this.props.vertical) return;
+			if (!this.props.detachedKnob || this.props.disabled || this.props.orientation === 'vertical') return;
 
 			this.moveKnobByPointer(ev.clientX);
 		}
@@ -507,7 +509,7 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 		}
 
 		handleActivate = () => {
-			const {'aria-valuetext': ariaValueText, detachedKnob, disabled, vertical} = this.props;
+			const {'aria-valuetext': ariaValueText, detachedKnob, disabled, orientation} = this.props;
 
 			if (disabled) return;
 
@@ -523,7 +525,7 @@ const SliderDecorator = hoc(defaultConfig, (config, Wrapped) => {
 
 				let valueText = ariaValueText != null ? ariaValueText : this.state.value;
 				if (active) {
-					valueText = vertical ? verticalHint : horizontalHint;
+					valueText = (orientation === 'vertical') ? verticalHint : horizontalHint;
 				}
 
 				this.setState({active, valueText});